Updated parameters class for new column and write
Updated the daesim2_analysis parameters class so that it can
also write the data to a JSON file (it can already read). This
also includes a new optional attribute (and column in the df) for
the optimised values of the parameters from a calibration routine.
To include this optional attribute/column, some of the parameters
class methods required updating.