Bug fix in update_and_run_model() function
The function used to update a model instance with new parameters and
then run the model given the new parameters had a bug when there
were phase-specific parameters (e.g. gdd_requirements). Each iteration
of the loop that updated the model instance with parameters that passed
through the phase-specific condition always updated the first phase
listed in the phase-specific parameters in param_info (pd.Dataframe()).
In my specific use case, germination was being updated every time with
every phase-specific parameter value, even if the actual param value was
for a different phase. This meant no subsequent phase listed in param_info
was actually being updated in the model instance.