Update forcing data for explicit time_discretisation
The forcing data class was setting nrundays_f in the
time_discretisation without explicitly knowing the end date. This
can lead to issues if there is poorly structured input forcing data
e.g. that has missing dates. It is safer to define the end date
with the DOY and Year columns and then determine time discretisation